pyproject.toml
[project]
name = "short-term-1-26"
version = "0.1.0"
description = "Dataset-backed web interface for an image detector workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch",
    "torchvision",
    "tqdm>=4.67.3",
]

[project.scripts]
detector-web = "main:main"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
